Encapsulate `ProjectionMapping` as a struct (#8033)
