+++ datafusion/physical-expr/src/utils.rs
@@ -362,15 +362,12 @@ pub fn merge_vectors(
 #[cfg(test)]
 mod tests {
     use std::fmt::{Display, Formatter};
-    use std::ops::Not;
     use std::sync::Arc;
 
     use super::*;
-    use crate::equivalence::EquivalenceProperties;
     use crate::expressions::{binary, cast, col, in_list, lit, Column, Literal};
     use crate::PhysicalSortExpr;
 
-    use arrow::compute::SortOptions;
     use arrow_array::Int32Array;
     use arrow_schema::{DataType, Field, Schema};
     use datafusion_common::cast::{as_boolean_array, as_int32_array};
@@ -631,232 +628,6 @@ mod tests {
         let result = as_boolean_array(&result)?;
 
         assert_eq!(&expected, result);
-        Ok(())
-    }
-
-    #[test]
-    fn test_get_indices_of_matching_sort_exprs_with_order_eq() -> Result<()> {
-        let sort_options = SortOptions::default();
-        let sort_options_not = SortOptions::default().not();
-
-        let schema = Schema::new(vec![
-            Field::new("a", DataType::Int32, true),
-            Field::new("b", DataType::Int32, true),
-        ]);
-        let col_a = &col("a", &schema)?;
-        let col_b = &col("b", &schema)?;
-        let required_columns = [col_b.clone(), col_a.clone()];
-        let mut eq_properties = EquivalenceProperties::new(Arc::new(schema));
-        eq_properties.add_new_orderings([vec![
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("b", 1)),
-                options: sort_options_not,
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("a", 0)),
-                options: sort_options,
-            },
-        ]]);
-        let (result, idxs) = eq_properties.find_longest_permutation(&required_columns);
-        assert_eq!(idxs, vec![0, 1]);
-        assert_eq!(
-            result,
-            vec![
-                PhysicalSortExpr {
-                    expr: col_b.clone(),
-                    options: sort_options_not
-                },
-                PhysicalSortExpr {
-                    expr: col_a.clone(),
-                    options: sort_options
-                }
-            ]
-        );
-
-        let schema = Schema::new(vec![
-            Field::new("a", DataType::Int32, true),
-            Field::new("b", DataType::Int32, true),
-            Field::new("c", DataType::Int32, true),
-        ]);
-        let col_a = &col("a", &schema)?;
-        let col_b = &col("b", &schema)?;
-        let required_columns = [col_b.clone(), col_a.clone()];
-        let mut eq_properties = EquivalenceProperties::new(Arc::new(schema));
-        eq_properties.add_new_orderings([
-            vec![PhysicalSortExpr {
-                expr: Arc::new(Column::new("c", 2)),
-                options: sort_options,
-            }],
-            vec![
-                PhysicalSortExpr {
-                    expr: Arc::new(Column::new("b", 1)),
-                    options: sort_options_not,
-                },
-                PhysicalSortExpr {
-                    expr: Arc::new(Column::new("a", 0)),
-                    options: sort_options,
-                },
-            ],
-        ]);
-        let (result, idxs) = eq_properties.find_longest_permutation(&required_columns);
-        assert_eq!(idxs, vec![0, 1]);
-        assert_eq!(
-            result,
-            vec![
-                PhysicalSortExpr {
-                    expr: col_b.clone(),
-                    options: sort_options_not
-                },
-                PhysicalSortExpr {
-                    expr: col_a.clone(),
-                    options: sort_options
-                }
-            ]
-        );
-
-        let required_columns = [
-            Arc::new(Column::new("b", 1)) as _,
-            Arc::new(Column::new("a", 0)) as _,
-        ];
-        let schema = Schema::new(vec![
-            Field::new("a", DataType::Int32, true),
-            Field::new("b", DataType::Int32, true),
-            Field::new("c", DataType::Int32, true),
-        ]);
-        let mut eq_properties = EquivalenceProperties::new(Arc::new(schema));
-
-        // not satisfied orders
-        eq_properties.add_new_orderings([vec![
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("b", 1)),
-                options: sort_options_not,
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("c", 2)),
-                options: sort_options,
-            },
-            PhysicalSortExpr {
-                expr: Arc::new(Column::new("a", 0)),
-                options: sort_options,
-            },
-        ]]);
-        let (_, idxs) = eq_properties.find_longest_permutation(&required_columns);
-        assert_eq!(idxs, vec![0]);
-
-        Ok(())
-    }
-
-    #[test]
-    fn test_normalize_ordering_equivalence_classes() -> Result<()> {
-        let sort_options = SortOptions::default();
-
-        let schema = Schema::new(vec![
-            Field::new("a", DataType::Int32, true),
-            Field::new("b", DataType::Int32, true),
-            Field::new("c", DataType::Int32, true),
-        ]);
-        let col_a_expr = col("a", &schema)?;
-        let col_b_expr = col("b", &schema)?;
-        let col_c_expr = col("c", &schema)?;
-        let mut eq_properties = EquivalenceProperties::new(Arc::new(schema.clone()));
-
-        eq_properties.add_equal_conditions(&col_a_expr, &col_c_expr);
-        let others = vec![
-            vec![PhysicalSortExpr {
-                expr: col_b_expr.clone(),
-                options: sort_options,
-            }],
-            vec![PhysicalSortExpr {
-                expr: col_c_expr.clone(),
-                options: sort_options,
-            }],
-        ];
-        eq_properties.add_new_orderings(others);
-
-        let mut expected_eqs = EquivalenceProperties::new(Arc::new(schema));
-        expected_eqs.add_new_orderings([
-            vec![PhysicalSortExpr {
-                expr: col_b_expr.clone(),
-                options: sort_options,
-            }],
-            vec![PhysicalSortExpr {
-                expr: col_c_expr.clone(),
-                options: sort_options,
-            }],
-        ]);
-
-        let oeq_class = eq_properties.oeq_class().clone();
-        let expected = expected_eqs.oeq_class();
-        assert!(oeq_class.eq(expected));
-
-        Ok(())
-    }
-
-    #[test]
-    fn project_empty_output_ordering() -> Result<()> {
-        let schema = Schema::new(vec![
-            Field::new("a", DataType::Int32, true),
-            Field::new("b", DataType::Int32, true),
-            Field::new("c", DataType::Int32, true),
-        ]);
-        let mut eq_properties = EquivalenceProperties::new(Arc::new(schema.clone()));
-        let ordering = vec![PhysicalSortExpr {
-            expr: Arc::new(Column::new("b", 1)),
-            options: SortOptions::default(),
-        }];
-        eq_properties.add_new_orderings([ordering]);
-        let projection_mapping = vec![
-            (
-                Arc::new(Column::new("b", 1)) as _,
-                Arc::new(Column::new("b_new", 0)) as _,
-            ),
-            (
-                Arc::new(Column::new("a", 0)) as _,
-                Arc::new(Column::new("a_new", 1)) as _,
-            ),
-        ];
-        let projection_schema = Arc::new(Schema::new(vec![
-            Field::new("b_new", DataType::Int32, true),
-            Field::new("a_new", DataType::Int32, true),
-        ]));
-        let orderings = eq_properties
-            .project(&projection_mapping, projection_schema)
-            .oeq_class()
-            .output_ordering()
-            .unwrap_or_default();
-
-        assert_eq!(
-            vec![PhysicalSortExpr {
-                expr: Arc::new(Column::new("b_new", 0)),
-                options: SortOptions::default(),
-            }],
-            orderings
-        );
-
-        let schema = Schema::new(vec![
-            Field::new("a", DataType::Int32, true),
-            Field::new("b", DataType::Int32, true),
-            Field::new("c", DataType::Int32, true),
-        ]);
-        let eq_properties = EquivalenceProperties::new(Arc::new(schema));
-        let projection_mapping = vec![
-            (
-                Arc::new(Column::new("c", 2)) as _,
-                Arc::new(Column::new("c_new", 0)) as _,
-            ),
-            (
-                Arc::new(Column::new("b", 1)) as _,
-                Arc::new(Column::new("b_new", 1)) as _,
-            ),
-        ];
-        let projection_schema = Arc::new(Schema::new(vec![
-            Field::new("c_new", DataType::Int32, true),
-            Field::new("b_new", DataType::Int32, true),
-        ]));
-        let projected = eq_properties.project(&projection_mapping, projection_schema);
-        // After projection there is no ordering.
-        assert!(projected.oeq_class().output_ordering().is_none());
-
         Ok(())
     }
 }
